+++ apps/ymir-sdl3/src/app/ui/widgets/settings_widgets.hpp
@@ -33,6 +33,7 @@ namespace settings::video {
 
         void Deinterlace(SharedContext &ctx);
         void TransparentMeshes(SharedContext &ctx);
+        void ResolutionScaling(SharedContext &ctx);
 
     } // namespace enhancements